move recurrenceComponent logic into own component
